requirements.txt
databricks-sql-connector
matplotlib
pandas
scikit_learn
seaborn
streamlit